Start of devlopment
I have added more detail into the floor plan, I have done this for all rooms, I have added jumping and sprinting ( No Jumping in the Apartment ). I have added a health system with a slider, I have made it so the player only regens health while inside of the apartment. (I have made several models to use in the apartment. It's not a lot but its a start as I need to work on wemos aswell so I just need to keep topping this up.