Blender/Audio
I have slightly changed the way the movement code works when in the air ( I still need to adjusy this as it feels very floaty ). I have created a simple fridge model and created some sand / desert terrain with a texture. I have created several sound that I plan on using ( I have only implemented one ), I have made the FOV chnage when near the TV more dramatic but when going out and back in it does not work properly. I have created a loading screen with a 5 second timer so the player is not shocked when going to the other area.